# go-runner/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A microservices management API built with Go and Gin",
        "title": "Go Runner API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/system/alerts": {
            "get": {
                "description": "Get system alerts with filtering options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get system alerts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alert type filter",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Alert level filter",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Active alerts only",
                        "name": "active",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "System alerts",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/system/config": {
            "get": {
                "description": "Get system monitoring configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get system configuration",
                "responses": {
                    "200": {
                        "description": "System configuration",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update system monitoring configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Update system configuration",
                "parameters": [
                    {
                        "description": "System configuration",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_system.SystemConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated configuration",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/system/dashboard": {
            "get": {
                "description": "Get system dashboard with overview information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get system dashboard",
                "responses": {
                    "200": {
                        "description": "System dashboard",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/system/info": {
            "get": {
                "description": "Get comprehensive system information including CPU, memory, disk, and network",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get system information",
                "responses": {
                    "200": {
                        "description": "System information",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/system/status": {
            "get": {
                "description": "Get current system status and health indicators",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get system status",
                "responses": {
                    "200": {
                        "description": "System status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_system.SystemConfig": {
            "type": "object",
            "properties": {
                "alert_email": {
                    "description": "Alert email address",
                    "type": "string"
                },
                "alert_webhook": {
                    "description": "Alert webhook URL",
                    "type": "string"
                },
                "check_interval": {
                    "description": "Check interval in seconds",
                    "type": "integer"
                },
                "cpu_limit": {
                    "description": "CPU usage threshold (%)",
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "disk_limit": {
                    "description": "Disk usage threshold (%)",
                    "type": "number"
                },
                "enable_alerts": {
                    "description": "Enable alerting",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "memory_limit": {
                    "description": "Memory usage threshold (%)",
                    "type": "number"
                },
                "network_limit": {
                    "description": "Network usage threshold (Mbps)",
                    "type": "number"
                },
                "retention_days": {
                    "description": "Metrics retention in days",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}